Fix #36, #41, 48
The source of the problem wasn't Firefox but I suppose it was Linux. On the
previous versions of Markdown Viewer, the converter from text/x-markdown to
text/plain worked but the mime type text/x-markdown doesn't exist (anymore?)
on Linux (see /usr/share/mime/). The correct mime type is in fact text/markdown